Code cleaned, realData array ready to be sent
