+++ core/server/app.js
@@ -11,12 +11,14 @@ var compression = require('compression');
 var bodyParser = require('body-parser');
 var net = require('net');
 var makeTcpReceiver = require('./makeTcpReceiver');
-var fs = require('fs');
 
 var database = require('../database');
+var datasetParser = require('./readDataset.js')
 
 var PORT = 4000;
 var DEBUG = process.env.NODE_ENV === "development" ? true : false;
+
+// var realDatas; // Set in the promise.all at the end of the file.
 
 var endpointConfig =
     {
@@ -134,70 +136,16 @@ server.listen(PORT, function () {
 
 // Load the real dataset.
 
-// remove accents from a String
-function RemoveAccents(strAccents) {
-    var strAccents = strAccents.split('');
-    var strAccentsOut = new Array();
-    var strAccentsLen = strAccents.length;
-    var accents = 'ÀÁÂÃÄÅàáâãäåÒÓÔÕÕÖØòóôõöøÈÉÊËèéêëðÇçÐÌÍÎÏìíîïÙÚÛÜùúûüÑñŠšŸÿýŽž';
-    var accentsOut = "AAAAAAaaaaaaOOOOOOOooooooEEEEeeeeeCcDIIIIiiiiUUUUuuuuNnSsYyyZz";
-    for (var y = 0; y < strAccentsLen; y++) {
-        if (accents.indexOf(strAccents[y]) != -1) {
-            strAccentsOut[y] = accentsOut.substr(accents.indexOf(strAccents[y]), 1);
-        } else
-            strAccentsOut[y] = strAccents[y];
-    }
-    strAccentsOut = strAccentsOut.join('');
-    return strAccentsOut;
-}
-
-// parse a .csv and return an array of objects
-function parseCSVSync(csv, format, callback) {
-    var file = csv.toString().split('\n');
-    var _format = file.shift().split(';');
-    if (format)
-        _format = format;
-    var objArray = file.map(function (line) {
-        if (line.length === 0)
-            return null;
-        var i = 0;
-        var obj = {};
-        line.split(';').forEach(function (field) {
-            obj[(_format[i]).toString()] = field;
-            ++i;
-        })
-        return obj;
-    })
-    if (callback && typeof(callback) == "function")
-        callback(null, objArray);
-    return objArray;
-}
-
-// return the whole dataset as an array of objects like this: [{place, date}, {place, date}]
-function readDataset() {
-    return new Promise(function (resolve, reject) {
-        fs.readFile('/6element/data/all_data.csv', function (err, file) {
-            if (err) reject(err);
-
-            var dataset = parseCSVSync(file, ['place', 'date', 'who', 'what', 'quantity', 'unit'])
-            .map(function (object) {
-                if (object)
-                    return {place: object.place.replace('ST ', 'Saint-').toLowerCase(), date: new Date(object.date)};
-                else
-                    return (null);
-            });
-            resolve(dataset);
-        })
-    });
-}
-
-var PCsv = readDataset()
+var PCsv = datasetParser.readDataset()
+.catch(function (err) {
+    console.log('ERROR in dataset parsing:', err)
+})
 
 // return all the places in the DB in order to create a map name->id
 var PPlaces = database.complexQueries.getAllPlacesInfos()
-
-var getPlaceMeasurements
-
+.catch(function (err) {
+    console.log('ERROR in database retrieving:', err)
+})
 
 Promise.all([PCsv, PPlaces])
 .then(function(results) {
@@ -208,13 +156,13 @@ Promise.all([PCsv, PPlaces])
     var nameToID = {};
 
     places.forEach(function (place) {
-        nameToID[RemoveAccents(place.name.toLowerCase())] = place.id;
+        nameToID[datasetParser.removeAccents(place.name.toLowerCase())] = place.id;
     })
 
     // Cut the dataset into little arrays by place ID
     var datasByID = {};
     dataset.forEach(function(data) {
-        if (!data) return ;
+        if (!data) return;
         if (!data.place || !nameToID[data.place] || datasByID[nameToID[data.place]] === undefined) {
             datasByID[nameToID[data.place]] = [];
         }
@@ -227,9 +175,11 @@ Promise.all([PCsv, PPlaces])
             return data1.getTime() - data2.getTime();
         })
     })
-
-    console.log(datasByID)
-});
+    //realDatas = datasByID;
+})
+.catch(function (err) {
+    console.log('ERROR :', err)
+})
 
 // In dev, there are usually no sensor pushing.
 // Let's simulate that part
